backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pymupdf==1.23.5
Pillow==10.0.0
